% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{nitrates_small}
\alias{nitrates_small}
\title{Daily nitrate concentrations at NEON river sites (June–July 2019)}
\format{
A tibble with 183 rows and 3 variables:
\describe{
\item{site}{NEON site code (ARIK, CARI, LEWI)}
\item{date}{Date of observation (YYYY-MM-DD, UTC)}
\item{nitrate_mgL}{Daily mean nitrate concentration (milligrams per litre, mg/L)}
}
}
\source{
National Ecological Observatory Network (NEON), product DP1.20033.001.
}
\usage{
nitrates_small
}
\description{
A dataset containing daily mean nitrate concentrations for three NEON
freshwater sites, used for the \code{nitrateApp} Shiny app.
}
\examples{
data("nitrates_small", package = "nitrateApp")
head(nitrates_small)
}
\keyword{datasets}
